refactor: move filtered functions in filteredFunctions.tsx
